Support Oracle q-Quote Syntax
Oracle supports a quote syntax for defining text that does not require escaping quote characters. It's usually used for defining inline multi-line queries.

Defining as sqlQuote type in syntax file because it seems not to work when it shares with the sqlString. I suspect there's some sort of conflict in those cases.

https://docs.oracle.com/cd/B19306_01/server.102/b14200/sql_elements003.htm#i42617